requirements.txt
flask
openai
langchain
langchain-community
langchain-chroma
langchain-openai
python-dotenv
bs4